Move translations from frontend to backend
Implements the logic to move translations from frontend
to backend.
In this case, moves translations for survey-questions
from frontend to backend.
 - In the Frontend, for the parts that are generated
automatically, the backend has stored the texts in the
different languages.
    * It is very important that the questions
have their ids.
    * Then a request is made to the Evaluation backend
in order to get the questions that it wants to be shown
by supplying the language that is currently being used.
That way, the Evaluation is always displayed in
the correctly selected language.
    * The Evaluation object comes from the backend
supplying all its information in the language that is
currently being used in the frontend.
 - When the language is changed in the application menu,
the setLanguage method:
    1-) Makes a call to the backend. Implements the logic
in SurveyService.getEvaluationById, then calls this method.
This method is also called in SurveyResolverService.
    2-) Informs the new language to be used,
e.g.: this._transServ.use('es').
 - The code-logic of the resolve method changed.
 - In the setLanguage method, within the subscribe
of this._surveyService.getEvaluationById, the fields
that were translated are assigned to the fields
of the this._surveyValueService._evaluation object.
 - Implements a new version of the deepcopy function
called cloneValue. This function is global, and is in
the core/utils/helpers.ts file of the toco-lib library.
 - Implements the unit tests for the new cloneValue function
in the core/test/utils/test-helpers.ts file.
 - Sets the correct values ​​for the objects:
evaluationEmpty_Spanish, evaluationEmpty_English,
evaluationExample_Spanish, evaluationExample_English.
 - Deletes the translation keys in the es.json
and en.json files. A copy of these keys is kept.
 - Removes the translation pipes from the code (| translate).